Bug fixes and modifications in light sensor API

+++ include/temp_sensor.h
@@ -5,7 +5,7 @@
 
 float temp_sensor_read(void);
 void temp_sensor_config_regwrite(uint16_t);
-uint32_t temp_sensor_config_regread(void);
+uint16_t temp_sensor_config_regread(void);
 void temp_sensor_sd(uint8_t);
 void temp_sensor_config_conversion_rate(uint8_t);
